Fix UTF8 encoding
PowerShell uses "UTF-8 with BOM" when specifying "-Encoding UTF8". To generate a "UTF-8" file without BOM, we need to use the .NET System.IO.StreamWriter class with System.Text.UTF8Encoding specifying false for the constructor parameter that controls whether or not to output a BOM.

This script creates a StreamWriter with UTF-8 encoding without BOM and writes the output to the file using this StreamWriter. This way, you can guarantee that the output file is in the format "UTF-8" and not "UTF-8 with BOM".